feat: refactored test files, updated workflows
