+++ .github/workflows/hono_teardown.yml
@@ -1,4 +1,4 @@
-name: PR Closed
+name: Hono Teardown
 
 on:
   pull_request:
@@ -8,14 +8,14 @@ env:
   NEON_PROJECT_ID: ${{ secrets.NEON_PROJECT_ID }}
   NEON_API_TOKEN: ${{ secrets.NEON_API_TOKEN }}
   CF_API_TOKEN: ${{ secrets.CF_API_TOKEN }}
-  DEPLOY_TAG: ${{ github.event.number }}-stage-shafa-hono
+  DEPLOY_TAG: shafa-hono-stage-${{ github.event.number }}
 
 jobs:
   neon_delete_branch:
     runs-on: ubuntu-latest
     steps:
       - name: Delete Neon Branch
-        uses: neondatabase/delete-branch-action@v3.1.3
+        uses: neondatabase/delete-branch-action@v3.1.4
         with:
           project_id: ${{ env.NEON_PROJECT_ID }}
           branch: ${{ env.DEPLOY_TAG }}
@@ -28,5 +28,5 @@ jobs:
         uses: cloudflare/wrangler-action@v3
         with:
           apiToken: ${{ env.CF_API_TOKEN }}
-          command: delete --name ${{ env.DEPLOY_TAG }} --env stage
+          command: delete --name ${{ env.DEPLOY_TAG }}
           workingDirectory: hono
