+++ hono/package.json
@@ -6,11 +6,11 @@
     "dev": "wrangler dev",
     "migrate": "tsx migrate.ts",
     "generate": "drizzle-kit generate:pg",
-    "test": "jest --silent --detectOpenHandles --testPathIgnorePatterns='neon.system' --testPathIgnorePatterns='remote.system' --testPathIgnorePatterns='/system'",
+    "test": "jest --silent --detectOpenHandles --testPathIgnorePatterns='neon.preflight' --testPathIgnorePatterns='remote.preflight' --testPathIgnorePatterns='/preflight'",
     "test:smoke": "jest --silent --testPathPattern=smoke.test.ts$",
-    "test:functional": "jest --silent --testPathPattern=functional.test.ts$",
+    "test:unit": "jest --silent --testPathPattern=unit.test.ts$",
     "test:integration": "jest --silent --testPathPattern=integration.test.ts$",
-    "test:system": "jest --silent local.system.test.ts",
+    "test:preflight": "jest --silent local.preflight.test.ts",
     "format": "prettier --log-level warn --write . && eslint --fix .",
     "check-ts": "tsc --noEmit --target esnext --project tsconfig.json",
     "check-prettier": "prettier --check .",
