+++ hono/test/integration/items_outfits.integration.test.ts
@@ -1,11 +1,11 @@
 import { drizzle } from 'drizzle-orm/postgres-js'
 import type { Context } from 'hono'
 import postgres from 'postgres'
-import * as schema from '../src/schema'
-import { clean, provision } from './utils/db'
+import * as schema from '../../src/schema'
+import { clean, provision } from '../utils/db'
 
 /**
- * Integration Tests
+ * Items to Outfits, Integration Tests
  *
  * These tests go beyond unit tests by testing the interaction between various
  * parts of the system as a group. They focus on how different functions work
@@ -14,13 +14,13 @@ import { clean, provision } from './utils/db'
  * tables/databases, etc.
  */
 
-const DB_NAME = 'integration_test'
+const DB_NAME = 'items_outfits_integration_test'
 
 // NOTE: Beware of jest hoisting!
 //       .mock() will be automatically hoisted to the top of the code block,
 //       because of this function decomposition is not possible without overhead
-jest.mock('../src/utils/injectDB', () => {
-  const originalModule = jest.requireActual('../src/utils/injectDB')
+jest.mock('../../src/utils/injectDB', () => {
+  const originalModule = jest.requireActual('../../src/utils/injectDB')
   return {
     __esModule: true,
     ...originalModule,
